pyproject.toml
[tool.poetry]
name = "backend-tfg"
packages = [{include = "flask_app"}]
version = "0.1.0"
description = ""
authors = ["José Carlos López Aguilar"]
readme = "README.md"

[tool.taskipy.tasks]
phpmyadmin = "sudo docker compose up phpmyadmin"
db = "sudo docker compose up db"
flask = "flask --app flask_app run --debug"
docker = "sudo docker compose --profile debug up"
pre_dev = "task docker -d"
dev = "sleep 5 && task flask"

[tool.poetry.dependencies]
python = ">3.9.0,<3.9.1 || >3.9.1,<4.0"
flask = "^3.1.0"
xlsxwriter = "^3.2.2"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
flask-mail = "^0.10.0"
flask-marshmallow = "^1.3.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
marshmallow-sqlalchemy = "^1.4.1"
cryptography = "^44.0.2"
argon2-cffi = "^23.1.0"
flask-jwt-extended = "^4.7.1"
flask-cors = "^5.0.1"
requests = "^2.32.4"


[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
